webrtcAPI/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==4.2.16
django-crispy-forms==2.3
django-encrypted-model-fields==0.6.5
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
pycparser==2.22
python-decouple==3.8
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
gevent
ws4py
whitenoise